pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-relay"
version = "0.1.0"
description = "Multi-agent workflow orchestrator — docker-compose for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Agent Relay Contributors" },
]
keywords = ["ai", "agents", "workflow", "orchestration", "cli", "tui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "typer>=0.12,<1.0",
    "rich>=13.0,<14.0",
    "textual>=0.40,<1.0",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "watchfiles>=0.21,<1.0",
]

[project.scripts]
relay = "relay.cli:app"

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
dev = ["pytest>=8.0", "pytest-cov", "ruff"]

[tool.hatch.build.targets.wheel]
packages = ["src/relay"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
